build Sailfish app with static plugins as for IOS
To get rid of warnings and double process on start I build the app in one
piece of cake. The generated binary is called 'noson'.

This change will work only with the latest sources of the app, because the
project configuration has been updated to manage the setting QT_STATICPLUGIN
for the Sailfish build.